% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_norm.R
\name{min_max_norm}
\alias{min_max_norm}
\title{Normalise a dataset from erasmusladder}
\usage{
min_max_norm(x)
}
\arguments{
\item{x}{The data (should be provided by erasmusladder)}
}
\value{

}
\description{
Normalise a dataset from erasmusladder
}
\examples{
grouped_dat <- data \%>\% select(columns, !!col_y, Session) \%>\%
filter(Session == session_num) \%>\% # Filter out the sessions called
pivot_longer(cols = columns, # Pivot the tibble to be tidy
            names_to = "step_type",
                        values_to = "values") \%>\%
                        group_by(step_type, !!col_y) # And group it to only get the variables we need
norm_dat <- min_max_norm(x = grouped_dat$values) \%>\%
as.data.frame(row.names = NULL, optional = TRUE)
}
