% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_graph.R
\name{time_graph}
\alias{time_graph}
\title{Make a graph based on data provided by the erasmusladder}
\usage{
time_graph(data, column, title, y_as)
}
\arguments{
\item{data}{a dataset, provided by the erasmusladder}

\item{column}{the name of the collum wich u want to use should be called with quo function}

\item{title}{how you want to name the graph}

\item{y_as}{how you want to name the y-axis}
}
\value{
a graph
}
\description{
Make a graph based on data provided by the erasmusladder
}
\examples{
data_time <- read_csv(here("data", "data.csv"))
time_graph(data_time, column = quo(`Trial duration(s)`), "Average trail duration per session per genotype", "Duration (s)")
}
