% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_violin_graph.R
\name{time_violin}
\alias{time_violin}
\title{Violion plotting for time graphs}
\usage{
time_violin(data, column, title, y_as, session_num)
}
\arguments{
\item{data}{a dataset provided by the erasmusladder}

\item{column}{wich collum you like to select, should be called with the quo function}

\item{title}{the title of your graph}

\item{y_as}{the name of the y-axis}

\item{session_num}{the amount of sessions you want to use}
}
\value{

}
\description{
Violion plotting for time graphs
}
\examples{
ruwe_data <- read_csv(here::here("data/data.csv"))
time_violin(dat_time, column = quo(`Trial duration(s)`), "Average trial duration per genotype", "Duration (s)", session_num = 1:20)
}
