% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_graph.R
\name{var_graph}
\alias{var_graph}
\title{This function function summaries data from custom columns}
\usage{
var_graph(data, col_steptype, col_y, session_num)
}
\arguments{
\item{data}{a dataset provided by R}

\item{col_steptype}{wich collum you like to select, should be called with the quo function}

\item{col_y}{The collum that contains the group you want to use, should be called with the quo function}

\item{session_num}{The amount of sessions you want to look at}
}
\value{

}
\description{
This function function summaries data from custom columns
}
\examples{
ruwe_data <- read_csv(here::here("data/data.csv"))
data_var_ruw <- ruwe_data \%>\% as.tibble() \%>\%
select("genotype", ends_with("Steps\%"), contains("Jumps\%"), `L - L Backsteps\%`, `H - L Long Steps\%`, "Session")
data_var_ruw$genotype <- as.factor(data_var_ruw$genotype)
var_graph(data_var_ruw, col_steptype = quo("H - L"), col_y = quo(genotype), session_num = 10:20)
}
